[rename & correct] [#100] name convention
